Made the creation of an empty case to display detailed information in a separate function.
(cherry picked from commit c1a837d7a76b22f6664b9353229d75b729fcbc1e)
