--- frontend/src/components/testcases/styles.testcases.css
@@ -1,46 +1,0 @@
-.mainGrid {
-    margin-top: 0;
-    position: absolute;
-    display: flex;
-    height: 91.5%;
-    width: 100%;
-}
-
-.leftGrid {
-    overflow-y: auto;
-    max-height: 100%;
-    width: 80%;
-}
-
-.rightGrid {
-    background-color: #eeeeee;
-    width: 20%;
-}
-
-.rightGridButtons {
-    text-align: center
-}
-
-.buttonCreateCase {
-    margin: 15px;
-    min-width: 70%;
-    height: 45%;
-    background-color: #FFFFFF;
-    color: #000000;
-}
-
-.buttonCreateSuite {
-    min-width: 70%;
-    height: 45%;
-    background-color: #be0909;
-    color: #FFFFFF;
-}
-
-.mainGridFolderStructure {
-    height: 67%
-}
-.tables {
-    padding: 9px 0 0 150px;
-    margin-right: 10px;
-    min-width: 300px
-}
