+++ frontend/src/components/testcases/suites.component.tsx
@@ -17,6 +17,7 @@ import PaginationSuitesComponent from "./pagination.suites.component";
 import {useParams, useNavigate} from "react-router-dom";
 import FormControl from "@mui/material/FormControl";
 import useStylesGlobal from "../../styles/styles";
+import {myCase} from "../models.interfaces";
 
 
 export const CustomWidthTooltip = styled(({className, ...props}: TooltipProps) => (
@@ -41,18 +42,6 @@ export const CustomWidthTooltip = styled(({className, ...props}: TooltipProps) =
         fontSize: 25,
     },
 }));
-
-export interface myCase {
-    id: number;
-    name: string;
-    suite: number;
-    scenario: string;
-    project: number;
-    estimate: number | null;
-    teardown: string;
-    setup: string;
-    url?: string;
-}
 
 export interface treeSuite {
     id: number;
@@ -81,18 +70,7 @@ const SuitesComponent = () => {
     const [infoSuiteForEdit, setInfoSuiteForEdit] = useState<{ id: number, name: string } | null>(null)
     const [lastEditCase, setLastEditCase] = useState<number>(-1)
     const [selectedSuiteCome, setSelectedSuiteCome] = useState<{ id: number, name: string } | null>(null)
-    const [detailedCaseInfo, setDetailedCaseInfo] = useState<{ show: boolean, myCase: myCase }>({
-        show: false, myCase: {
-            id: -1,
-            name: "",
-            suite: -1,
-            scenario: "",
-            project: -1,
-            setup: "",
-            teardown: "",
-            estimate: -1
-        }
-    })
+    const [detailedCaseInfo, setDetailedCaseInfo] = useState<{ show: boolean, myCase: myCase }>(SuiteCaseService.getEmptyDetailedCaseInfo())
     const {selectedSuiteId} = useParams()
     const [selectedSuiteForTreeView, setSelectedSuiteForTreeView] = useState<treeSuite | undefined>(undefined)
     const navigate = useNavigate()
@@ -178,34 +156,34 @@ const SuitesComponent = () => {
                                                                         setTreeSuites={setTreeSuites}
                 />}
                 {selectedSuiteForTreeView === undefined &&
-                    <PaginationSuitesComponent countOfSuitesOnPage={countOfSuitesOnPage} treeSuites={treeSuites}/>}
+                <PaginationSuitesComponent countOfSuitesOnPage={countOfSuitesOnPage} treeSuites={treeSuites}/>}
             </div>
             <div className={classes.rightGrid}>
                 <div className={classes.rightGridButtons}>
                     {selectedSuiteForTreeView !== undefined &&
-                        <div>
-                            <Button sx={{
-                                margin: "15px 15px 0 15px",
-                                minWidth: "70%",
-                                height: "45%",
-                                backgroundColor: "#FFFFFF",
-                                color: "#000000",
-                                "&:hover": {
-                                    backgroundColor: "#fffafa",
-                                }
-                            }} onClick={handleShowCreationCase}>Создать
-                                тест-кейс</Button>
-                            <CreationCase show={showCreationCase} setShow={setShowCreationCase}
-                                          selectedSuiteCome={selectedSuiteCome} setTreeSuites={setTreeSuites}
-                                          infoCaseForEdit={infoCaseForEdit}
-                                          setInfoCaseForEdit={setInfoCaseForEdit}
-                                          setDetailedCaseInfo={setDetailedCaseInfo}
-                                          detailedCaseInfo={detailedCaseInfo}
-                                          setLastEditCase={setLastEditCase}
-                                          setSelectedSuiteForTreeView={setSelectedSuiteForTreeView}
-                                          selectedSuiteForTreeView={selectedSuiteForTreeView}
-                            />
-                        </div>}
+                    <div>
+                        <Button sx={{
+                            margin: "15px 15px 0 15px",
+                            minWidth: "70%",
+                            height: "45%",
+                            backgroundColor: "#FFFFFF",
+                            color: "#000000",
+                            "&:hover": {
+                                backgroundColor: "#fffafa",
+                            }
+                        }} onClick={handleShowCreationCase}>Создать
+                            тест-кейс</Button>
+                        <CreationCase show={showCreationCase} setShow={setShowCreationCase}
+                                      selectedSuiteCome={selectedSuiteCome} setTreeSuites={setTreeSuites}
+                                      infoCaseForEdit={infoCaseForEdit}
+                                      setInfoCaseForEdit={setInfoCaseForEdit}
+                                      setDetailedCaseInfo={setDetailedCaseInfo}
+                                      detailedCaseInfo={detailedCaseInfo}
+                                      setLastEditCase={setLastEditCase}
+                                      setSelectedSuiteForTreeView={setSelectedSuiteForTreeView}
+                                      selectedSuiteForTreeView={selectedSuiteForTreeView}
+                        />
+                    </div>}
                     <Button sx={{
                         marginTop: "15px",
                         minWidth: "70%",
@@ -227,26 +205,26 @@ const SuitesComponent = () => {
                     />
                 </div>
                 {selectedSuiteForTreeView === undefined &&
-                    <div>
-                        <FormControl sx={{minWidth: "90%", margin: "25px 0px 0px 15px"}}
-                                     className={classesGlobal.textFieldSelectCreationCaseSuite}>
-                            <InputLabel>Количество сьют на одной странице</InputLabel>
-                            <Select
-                                value={countOfSuitesOnPage}
-                                label="Количество сьют на одной странице"
-                                onChange={(e) => onChangeSuitesOnPage(e)}
-                                MenuProps={MenuProps}
-                            >
-                                {Array.from({length: (stop - start) / step + 1}, (_, i) => start + i * step).map((num, index) =>
-                                    <MenuItem key={index}
-                                              value={num}>{num}</MenuItem>)}
-                            </Select>
-                        </FormControl>
-                    </div>}
+                <div>
+                    <FormControl sx={{minWidth: "90%", margin: "25px 0px 0px 15px"}}
+                                 className={classesGlobal.textFieldSelectCreationCaseSuite}>
+                        <InputLabel>Количество сьют на одной странице</InputLabel>
+                        <Select
+                            value={countOfSuitesOnPage}
+                            label="Количество сьют на одной странице"
+                            onChange={(e) => onChangeSuitesOnPage(e)}
+                            MenuProps={MenuProps}
+                        >
+                            {Array.from({length: (stop - start) / step + 1}, (_, i) => start + i * step).map((num, index) =>
+                                <MenuItem key={index}
+                                          value={num}>{num}</MenuItem>)}
+                        </Select>
+                    </FormControl>
+                </div>}
                 {selectedSuiteForTreeView !== undefined &&
-                    <div className={classes.mainGridFolderStructure}>
-                        {memoizedValueFolderStructureOfSuites}
-                    </div>}
+                <div className={classes.mainGridFolderStructure}>
+                    {memoizedValueFolderStructureOfSuites}
+                </div>}
             </div>
         </div>
     )
